Import plugins with global visibility
The default behavior of DL::OpenLibrary/dlopen that we use to load plugins is
to load the symbols using RTLD_LOCAL on Linux which causes issues in python
plugins. In particular, if we load a python plugin from C++, the plugin
interpreter that starts in python_plugins.so hides symbols of libpython.so
behind RTDL_LOCAL making them inaccessible to Python modules (i.e. numpy)
that a python plugin might use.

Co-authored-by: Ben Boeckel <ben.boeckel@kitware.com>