cmake: make preprocessor standard conforming on windows
In the core of the macros defined in pluggable_macros.h is the IS_PROBE macro
which evaluates wrongly with msvc. In particular in the snippet below:

PROBE: IS_PROBE(PROBE())

the output is PROBE: 0 while it should be PROBE: 1.
Indeed gcc -E snippet.c gives the right anwser.

In the traditional preprocessor, if a macro forwards one of its arguments to
another dependent macro then the argument doesn't get "unpacked" when it's
inserted[1].  Adding the flag /Zc:preprocessor makes the preprocessor standard
complient and evaluates the above snippet to 1 as expected.
See also [2]:

[1] https://learn.microsoft.com/en-us/cpp/preprocessor/preprocessor-experimental-overview?view=msvc-170#macro-arguments-are-unpacked
[2] https://learn.microsoft.com/en-us/cpp/preprocessor/preprocessor-experimental-overview