Fix `tempnam` warning from gcc
gcc produced this warning on linking:
warning: the use of `tempnam' is dangerous, better use `mkstemp'

Since this method is only used in testing, we don't worry about
closing the file and reopening it later. We also add suffix for
file extension typing and assume it's ok to do so.