python: calculate PYTHONHOME based on linked library on windows
Activating a virtual environment prior to unsets PYTHONHOME and thus when
PyInitialize is called in plugins_from_python plugin it cannot find system
python packages. To resolve the issue we use the same technique  as in VTK: we
evaluate PYTHONHOME based on the python library binary that the plugin is
linked to.