python: support externally running castxml for MSVC
Due to https://github.com/CastXML/pygccxml/issues/180
we cannot generate the python wrappings on windows using MSVC and C++17.
In the meantime we apply the compose the castxml command ourselves and run it
as part of the script.  At this point pygccxml is used only to parse the generated xml
file.