cmake: fix python install path
Previously python packages would be placed in wrong places outside `CMAKE_INSTALL_PREFIX`.

If `library_dir` is unset during the call of `kwiver_add_library`,
it gets the (correct) default value. However, setting it to an empty string, it
is used as a prefix for the install path.
This causes the install path to be /lib/python3.8/site_packages instead of
${CMAKE_INSTALL_PREFIX}/lib/python3.7/site_packages.